Create innertext.html
A basic javascript code to show the basic use of javascript in web development. Kindly add as this shows one the most common uses of JS, that is changing text in HTML5 DIV Elements according to a certain function or variable. This module also shows the use of 'onclick' function, another commonly used function for button click events.