Create styleswithjs.html
This shows how we can integrate CSS styles with JS functions. This is very useful for website color themes like Dark/Light which is nowadays a commonly available feature on many websites for night reading as well as lesser light emitting by the screen. It also shows how the ID attribute can be used to control styles with both CSS and JS.